address bchocho comments
