+++ consensus/src/quorum_store/tests/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod batch_generator_test;
+mod batch_proof_queue_test;
 mod batch_requester_test;
 mod batch_store_test;
 mod direct_mempool_quorum_store_test;
@@ -9,4 +10,3 @@ mod proof_coordinator_test;
 mod proof_manager_test;
 mod quorum_store_db_test;
 mod types_test;
-mod utils;
